+++ data/storage.txt
@@ -1,5 +1,6 @@
 T | 0 | more work
-T | 0 | get a life
 D | 1 | call the police | by:2023-02-03T20:00
 D | 0 | call the ambulance | by:2024-02-02T08:00
 E | 0 | court hearing | from:2023-08-01T18:00 | to:2023-08-01T20:00
+T | 0 | go and do some stuff
+E | 1 | class | from:2023-01-25T19:00 | to:2023-01-25T20:00
