--- src/main/java/InvalidStorageException.java
@@ -1,5 +1,0 @@
-public class InvalidStorageException extends Exception{
-    InvalidStorageException(String msg) {
-        super(msg);
-    }
-}
