+++ src/main/java/chatbot/Parser.java
@@ -1,10 +1,31 @@
+package chatbot;
+
+import command.AddCommand;
+import command.Command;
+import command.CommandNotFoundException;
+import command.DeleteCommand;
+import command.ExitCommand;
+import command.ListCommand;
+import command.MarkAsDoneCommand;
+import command.MarkAsUndoneCommand;
+import static task.TaskType.inputToTask;
+
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.List;
 
 public class Parser {
-    protected static Map<String,String> splitArgs(String input) {
+
+    /**Splits a String line of input into its command (the first word), and its subsequent arguments.
+     * By default, the key "Command" maps to the name of the command input (the first word).
+     * The name of the Command is also a key to its default arguments; for eg: "delete 1" creates a mapping of "delete" -> "1".
+     * In the case of adding a new event, the name of the command is mapped to its description; for eg: "event" -> "attend wedding"
+     *
+     * @param input arguments to be split
+     * @return map a mapping of each argument name to its value
+     */
+    private static Map<String,String> splitArgs(String input) {
         // returns a Map mapping name of arg to arg value based on the raw format
         // raw format according to task is /<name> <value>
 
@@ -45,10 +66,10 @@ public class Parser {
             case "bye":
                 return new ExitCommand();
             default:
-                if (TaskType.inputToTask.containsKey(command)) { // check if command is to add a task
+                if (inputToTask.containsKey(command)) { // check if command is to add a task
                     String desc = argValues.get(command);
-                    argValues.put("Description", desc);
-                    return new AddCommand(TaskType.inputToTask.get(command), argValues);
+                    argValues.put("Description", desc); // change the for description from the name of Command
+                    return new AddCommand(inputToTask.get(command), argValues);
                 }
                 throw new CommandNotFoundException("I'm sorry, I don't recognise this command :/");
         }
