+++ src/main/java/command/Command.java
@@ -1,9 +1,15 @@
+package command;
+
+import chatbot.Storage;
+import chatbot.Ui;
+import task.TaskList;
+
 public abstract class Command {
 
     protected boolean isExit;
     protected boolean isSave;
 
-    Command() {
+    protected Command() {
     }
     public abstract void execute(TaskList tl, Ui ui, Storage storage) throws Exception;
