+++ src/main/java/command/ExitCommand.java
@@ -1,6 +1,12 @@
+package command;
+
+import chatbot.Storage;
+import chatbot.Ui;
+import task.TaskList;
+
 public class ExitCommand extends Command {
 
-    ExitCommand() {
+    public ExitCommand() {
         super();
         this.isExit = true;
         this.isSave = false;
