+++ src/main/java/command/MarkAsDoneCommand.java
@@ -1,22 +1,28 @@
-public class MarkAsUndoneCommand extends Command {
-    private String unmarkAtIndex;
+package command;
 
-    MarkAsUndoneCommand(String index) {
+import chatbot.Storage;
+import chatbot.Ui;
+import task.TaskList;
+
+public class MarkAsDoneCommand extends Command {
+    private String markAtIndex;
+
+    public MarkAsDoneCommand(String index) {
         this.isExit = false;
         this.isSave = true;
-        this.unmarkAtIndex = index;
+        this.markAtIndex = index;
     }
 
     @Override
     public void execute(TaskList tl, Ui ui, Storage storage) throws Exception {
         boolean alreadyMarked;
         try {
-            int i = Integer.valueOf(this.unmarkAtIndex) - 1;
-            alreadyMarked = tl.unmarkDone(i);
+            int i = Integer.valueOf(this.markAtIndex) - 1;
+            alreadyMarked = tl.markAsDone(i);
             if (alreadyMarked) {
-                ui.reply(tl.getTask(i).getDesc() + " is already undone!");
+                ui.reply(tl.getTask(i).getDesc() + " already marked as done!");
             } else {
-                ui.reply("Unmarked " + tl.getTask(i).getDesc() + ".");
+                ui.reply("Marked " + tl.getTask(i).getDesc() + " as done!");
             }
         } catch (NumberFormatException e) {
             throw new Exception("Please specify the task by its index number.");
