+++ src/main/java/task/Deadline.java
@@ -1,10 +1,12 @@
+package task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
 import java.util.Objects;
 
 public class Deadline extends Task {
-    protected LocalDateTime deadline;
+    private LocalDateTime deadline;
 
     Deadline(String description, String deadline) throws InvalidDateTimeException {
         super(description);
