+++ src/main/java/task/InvalidInputException.java
@@ -1,3 +1,5 @@
+package task;
+
 public class InvalidInputException extends Exception {
 
     InvalidInputException(String msg) {
