+++ src/main/java/task/TaskType.java
@@ -1,3 +1,5 @@
+package task;
+
 import java.util.HashMap;
 import java.util.Map;
